+++ plugin_main/plugin_export.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 	"runtime"
 	"runtime/debug"
+	"strings"
 	"sync"
 	"time"
 	"unsafe"
@@ -27,6 +28,7 @@ import (
 	"github.com/alibaba/ilogtail/pkg/config"
 	"github.com/alibaba/ilogtail/pkg/helper"
 	"github.com/alibaba/ilogtail/pkg/logger"
+	"github.com/alibaba/ilogtail/pkg/pipeline"
 	"github.com/alibaba/ilogtail/pkg/util"
 	"github.com/alibaba/ilogtail/plugin_main/flags"
 	"github.com/alibaba/ilogtail/pluginmanager"
@@ -262,6 +264,19 @@ func GetContainerMeta(containerID string) *C.struct_containerMeta {
 		}
 	}
 	return returnStruct
+}
+
+//export GetGoPlugins
+func GetGoPlugins() *C.char {
+	var buffer strings.Builder
+	helper.ConvertCreatorMapToString(&buffer, pipeline.MetricInputs)
+	helper.ConvertCreatorMapToString(&buffer, pipeline.ServiceInputs)
+	helper.ConvertCreatorMapToString(&buffer, pipeline.Processors)
+	helper.ConvertCreatorMapToString(&buffer, pipeline.Aggregators)
+	helper.ConvertCreatorMapToString(&buffer, pipeline.Flushers)
+	helper.ConvertCreatorMapToString(&buffer, pipeline.Extensions)
+	res := buffer.String()
+	return C.CString(res)
 }
 
 func initPluginBase(cfgStr string) int {
