+++ plugin_main/plugin_main_test.go
@@ -20,7 +20,6 @@ package main
 import (
 	"context"
 	"fmt"
-	"strings"
 	"testing"
 	"time"
 
@@ -30,7 +29,6 @@ import (
 	"github.com/alibaba/ilogtail/pkg/protocol"
 	"github.com/alibaba/ilogtail/pluginmanager"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -203,11 +201,4 @@ func TestSlowConfigWhenStop(t *testing.T) {
 	HoldOn(1)
 	time.Sleep(time.Second * 6)
 	require.Empty(t, pluginmanager.DisabledLogtailConfig)
-}
-
-func TestPluginsNames(t *testing.T) {
-	pluginNames := GetGoPlugins()
-	if len(pluginNames) > 0 {
-		assert.True(t, strings.Contains(pluginNames, "|"))
-	}
 }
